% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{recm-methods}
\alias{recm-methods}
\alias{coef.recm}
\alias{vcov.recm}
\alias{residuals.recm}
\alias{fitted.recm}
\alias{nobs.recm}
\alias{logLik.recm}
\alias{print.recm}
\alias{summary.recm}
\alias{print.summary.recm}
\title{Methods for fitted rational error correction models}
\usage{
\method{coef}{recm}(object, ...)

\method{vcov}{recm}(object, ...)

\method{residuals}{recm}(object, ...)

\method{fitted}{recm}(object, ...)

\method{nobs}{recm}(object, ...)

\method{logLik}{recm}(object, ...)

\method{print}{recm}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{recm}(object, ...)

\method{print}{summary.recm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object, x}{An object of class `"recm"` returned by [recm()].}

\item{...}{Ignored, present for consistency with the generics.}

\item{digits}{Number of significant digits used when printing.}
}
\description{
Methods for fitted rational error correction models
}
